UI部分公共出来
