+++ CT.WPF.MagicEffects.Demo/Themes/Generic.xaml
@@ -1,4 +1,4 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
+<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
     <SolidColorBrush x:Key="DefaultThemeColor" Color="#FF2D9F87" />
     <SolidColorBrush x:Key="DefaultBorderBrush" Color="Beige" />
     <SolidColorBrush x:Key="DefaultBrush" Color="BlueViolet" />
