--- CT.WPF.MagicEffects.Demo/UserControls/Common/CommonEffectControl.xaml
@@ -1,12 +1,0 @@
-﻿<UserControl
-    x:Class="CT.WPF.MagicEffects.Demo.UserControls.Common.CommonEffectControl"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="clr-namespace:CT.WPF.MagicEffects.Demo.UserControls.Common"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    d:DesignHeight="450"
-    d:DesignWidth="800"
-    mc:Ignorable="d">
-    <Grid />
-</UserControl>
