--- CT.WPF.MagicEffects.Demo/UserControls/Common/CommonEffectControl.xaml.cs
@@ -1,27 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows;
-using System.Windows.Controls;
-using System.Windows.Data;
-using System.Windows.Documents;
-using System.Windows.Input;
-using System.Windows.Media;
-using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
-using System.Windows.Shapes;
-
-namespace CT.WPF.MagicEffects.Demo.UserControls.Common {
-    /// <summary>
-    /// CommonEffectControl.xaml 的交互逻辑
-    /// </summary>
-    public partial class CommonEffectControl : UserControl {
-        public CommonEffectControl() {
-            InitializeComponent();
-        }
-        
-
-    }
-}
