+++ CT.WPF.MagicEffects.Demo/UserControls/MonochromeEffectControl.xaml
@@ -4,6 +4,7 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:Skin="https://github.com/aduskin"
     xmlns:behavior="http://schemas.microsoft.com/xaml/behaviors"
+    xmlns:common="clr-namespace:CT.WPF.MagicEffects.Demo.UserControls.Common"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:effects="clr-namespace:CT.WPF.MagicEffects;assembly=CT.WPF.MagicEffects"
     xmlns:local="clr-namespace:CT.WPF.MagicEffects.Demo.UserControls"
@@ -16,31 +17,8 @@
     <UserControl.DataContext>
         <vm:MonochromeEffectViewModel />
     </UserControl.DataContext>
-    <Grid Background="{DynamicResource DefaultBackground}">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="250" />
-            <ColumnDefinition Width="*" />
-        </Grid.ColumnDefinitions>
-        <!--  展示列表展示多个路径效果xmal  -->
-        <StackPanel Orientation="Vertical">
-            <Border
-                Height="100"
-                MinWidth="100"
-                VerticalAlignment="Top"
-                Style="{StaticResource ShowControlBorder}">
-                <Skin:AduUpload
-                    BorderBrush="{Binding RelativeSource={RelativeSource AncestorType={x:Type Skin:MetroWindow}}, Path=BorderBrush, Mode=TwoWay}"
-                    Cursor="Hand"
-                    Filter="图片文件|*.jpg;*.jpeg;*.png;*.bmp;*.gif"
-                    Foreground="{Binding RelativeSource={RelativeSource AncestorType={x:Type Skin:MetroWindow}}, Path=BorderBrush, Mode=TwoWay}"
-                    MultiSelect="False">
-                    <behavior:Interaction.Triggers>
-                        <behavior:EventTrigger EventName="FileAduUpload">
-                            <behavior:InvokeCommandAction Command="{Binding SelectImageCommand}" PassEventArgsToCommand="True" />
-                        </behavior:EventTrigger>
-                    </behavior:Interaction.Triggers>
-                </Skin:AduUpload>
-            </Border>
+    <common:CommonEffectControl>
+        <common:CommonEffectControl.Option>
             <Skin:AduTabControl>
                 <Skin:AduTabControl.Resources>
                     <Style TargetType="TabPanel">
@@ -65,7 +43,7 @@
                                                 Width="99"
                                                 Height="88"
                                                 Margin="2">
-                                                <Image Source="{Binding DataContext.SelectedImagePath, RelativeSource={RelativeSource AncestorType={x:Type UserControl}}}">
+                                                <Image Source="{Binding Path=Main.SelectedImagePath, Source={StaticResource Locator}}">
                                                     <Image.Effect>
                                                         <effects:MonochromeEffect FilterColor="{Binding FilterColor}" />
                                                     </Image.Effect>
@@ -96,37 +74,18 @@
                         MinWidth="50"
                         MinHeight="50"
                         Padding="5"
-                        ColorChange="ColorPicker_ColorChange"
                         DefaultColor="Transparent">
-                        <!--<behavior:Interaction.Triggers>
+                        <behavior:Interaction.Triggers>
                             <behavior:EventTrigger EventName="ColorChanged">
-                                <behavior:InvokeCommandAction Command="{Binding ChangeFilterColorCommand}"
-                                                              CommandParameter="{Binding RelativeSource={RelativeSource AncestorType={x:Type Skin:MetroColorPicker}}, Path=Background, Mode=TwoWay}" />
+                                <behavior:InvokeCommandAction Command="{Binding ChangeFilterColorCommand}" CommandParameter="{Binding ElementName=ColorPicker, Path=Background}" />
                             </behavior:EventTrigger>
-                        </behavior:Interaction.Triggers>-->
+                        </behavior:Interaction.Triggers>
                     </Skin:MetroColorPicker>
                 </Skin:AduTabItem>
             </Skin:AduTabControl>
-
-        </StackPanel>
-
-        <Skin:MetroScrollViewer Grid.Column="1" HorizontalScrollBarVisibility="Disabled">
-            <Grid>
-                <ScrollViewer
-                    Grid.Row="1"
-                    HorizontalAlignment="Center"
-                    VerticalAlignment="Bottom"
-                    HorizontalScrollBarVisibility="Auto"
-                    VerticalScrollBarVisibility="Auto">
-                    <Viewbox Height="700" Stretch="Uniform">
-                        <Image Source="{Binding SelectedImagePath}">
-                            <Image.Effect>
-                                <effects:MonochromeEffect FilterColor="{Binding SelectedEffectOption.FilterColor}" />
-                            </Image.Effect>
-                        </Image>
-                    </Viewbox>
-                </ScrollViewer>
-            </Grid>
-        </Skin:MetroScrollViewer>
-    </Grid>
+        </common:CommonEffectControl.Option>
+        <common:CommonEffectControl.ImageEffect>
+            <effects:MonochromeEffect FilterColor="{Binding SelectedEffectOption.FilterColor}" />
+        </common:CommonEffectControl.ImageEffect>
+    </common:CommonEffectControl>
 </UserControl>
